Refactor homepage to SSR deals and reuse client shell
